requirements.txt
numpy==1.21.5
pandas==1.4.4
Pillow==9.3.0
streamlit==1.16.0
